+++ NEWS.md
@@ -7,8 +7,8 @@ Update UBR and take Windows 11 26H1 into account.
 New presets/rules:
 
 * 2026/03/10 - `UnsetDebugPrograms`/`SetDebugPrograms`/`ViewDebugPrograms` -> Unset Debug Programs user right only to the Administrators group (Set use `$Global:SWMB_Custom` key `DebugPrograms`)
-* 2026/03/10 - `DisableTCBPrivilege`/`EnableTCBPrivilege`/`ViewTCBPrivilege` -> Disable Trusted Computing Base Privilege (Enable use `$Global:SWMB_Custom` key `TCBPrivilege`)
-* 2026/03/10 - `DisableCreateTokenObject`/`EnableCreateTokenObject`/`ViewCreateTokenObject` -> Disable Create a Token Object all accounts (Enable use `$Global:SWMB_Custom` key `CreateTokenObject` empty by default)
+* 2026/03/10 - `UnsetTCBPrivilege`/`SetTCBPrivilege`/`ViewTCBPrivilege` -> Unset Trusted Computing Base Privilege (Set use `$Global:SWMB_Custom` key `TCBPrivilege`)
+* 2026/03/10 - `UnsetCreateTokenObject`/`SetCreateTokenObject`/`ViewCreateTokenObject` -> Unset Create a Token Object all accounts (Set use `$Global:SWMB_Custom` key `CreateTokenObject` empty by default)
 * 2026/03/10 - `DisableAAnonymousSAMEnumeration`/`EnableAAnonymousSAMEnumeration`/`ViewAAnonymousSAMEnumeration` -> Disable Anonymous enumeration of SAM accounts
 * 2026/03/10 - `DisableAnonymousShareEnumeration`/`EnableAnonymousShareEnumeration`/`ViewAnonymousShareEnumeration` -> Disable Anonymous enumeration of shares
 * 2026/03/10 - `DisableAnonymousNameTranslation`/`EnableAnonymousNameTranslation`/`ViewAnonymousNameTranslation` -> Disable Anonymous SID/Name translation